requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
fastapi>=0.110.0
uvicorn>=0.27.0
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
beautifulsoup4>=4.12.0
matplotlib>=3.8.0
plotly>=5.18.0
python-telegram-bot>=21.0
websockets>=15.0.0
aiofiles>=23.0.0
feedparser>=6.0.0
google-genai>=1.0.0
youtube-search-python>=1.6.6
youtube-transcript-api>=0.6.0
# v1.0 雲端 LLM 模組
openai>=1.0.0
pyyaml>=6.0.0
# v0.3 Singer Agent 圖片處理
rembg>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0